--- stdlib/filepath/src/main/java/org/ballerinalang/stdlib/filepath/nativeimpl/Absolute.java
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.stdlib.filepath.nativeimpl;
-
-import org.ballerinalang.stdlib.filepath.Constants;
-import org.ballerinalang.stdlib.filepath.Utils;
-
-import java.nio.file.FileSystems;
-import java.nio.file.InvalidPathException;
-
-/**
- * The native class to get absolute value of the path.
- *
- * @since 0.995.0
- */
-public class Absolute {
-
-    public static Object externAbsolute(String inputPath) {
-        try {
-            String absolutePath = FileSystems.getDefault().getPath(inputPath).toAbsolutePath().toString();
-            return  absolutePath;
-        } catch (InvalidPathException ex) {
-            return Utils.getPathError(Constants.INVALID_PATH_ERROR, "Invalid path " + inputPath);
-        }
-    }
-
-}
